+++ weather-core/src/main/kotlin/me/fernando/weather/domain/LocationMother.kt
@@ -8,7 +8,7 @@ open class LocationMother {
         latitude: Double? = 40.4167,
         longitude: Double? = -3.7033,
         population: Long? = 1000000,
-        timezone: Long? = 7200,
+        timezone: Int? = 7200,
         sunrise: Long? = 1652677343,
         sunset: Long? = 1652728918,
     ): Location {
