+++ weather-delivery/src/main/kotlin/me/fernando/weather/api/dto/CityDto.kt
@@ -8,7 +8,7 @@ data class CityDto(
     @JsonProperty("country") val country: String, // Country code (GB, JP etc.)
     @JsonProperty("coord") val coordinate: CoordinateDto, // Coordinate of the city
     @JsonProperty("population") val population: Long, // Population of the city
-    @JsonProperty("timezone") val timezone: Long, // Timezone of the city
+    @JsonProperty("timezone") val timezone: Int, // Timezone of the city
     @JsonProperty("sunrise") val sunrise: Long, // Sunrise time
     @JsonProperty("sunset") val sunset: Long // Sunset time
 )
